MSVC warning fix
git-svn-id: https://svn.osgeo.org/gdal/trunk@29136 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/gcore/mdreader/reader_digital_globe.cpp
@@ -147,7 +147,7 @@ void GDALMDReaderDigitalGlobe::LoadMetadata()
                                                   "IMAGE.CLOUDCOVER");
     if(NULL != pszCloudCover)
     {
-        float fCC = CPLAtofM(pszCloudCover);
+        double fCC = CPLAtofM(pszCloudCover);
         if(fCC < 0)
         {
             m_papszIMAGERYMD = CSLAddNameValue(m_papszIMAGERYMD, MD_NAME_CLOUDCOVER,
@@ -164,7 +164,7 @@ void GDALMDReaderDigitalGlobe::LoadMetadata()
         pszCloudCover = CSLFetchNameValue(m_papszIMDMD, "IMAGE_1.cloudCover");
         if(NULL != pszCloudCover)
         {
-            float fCC = CPLAtofM(pszCloudCover);
+            double fCC = CPLAtofM(pszCloudCover);
             if(fCC < 0)
             {
                 m_papszIMAGERYMD = CSLAddNameValue(m_papszIMAGERYMD, MD_NAME_CLOUDCOVER,
